Fix test to have the same time precision as the signal-cli

+++ app/bot/chat_client/clients/signal/formaters/jsonrpc.py
@@ -303,9 +303,9 @@ class JsonRpcFormater(MessageFormater):
                 )
         return attachments
 
-    def _get_datetime_from_timestamp(self, timestamp: float) -> datetime:
+    def _get_datetime_from_timestamp(self, timestamp: int) -> datetime:
         # Receiving timestamp in millisec transform to sec
-        return datetime.fromtimestamp(timestamp / 1000.0)
+        return datetime.fromtimestamp(timestamp / 1000)
 
     def _get_timestamp_from_datetime(self, time: datetime) -> int:
         # Transform timestamp back to millisec
